% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_estaciones.R
\name{mapa_estaciones}
\alias{mapa_estaciones}
\title{Generar un mapa interactivo de las estaciones de monitoreo en Quito}
\usage{
mapa_estaciones()
}
\value{
Un objeto de mapa interactivo (leaflet) con las estaciones marcadas.
}
\description{
Esta función crea un mapa interactivo con las ubicaciones de las estaciones de monitoreo
de calidad del aire en Quito, utilizando la librería leaflet. Las estaciones vigentes
se muestran con marcadores verdes.
}
\examples{
\dontrun{
# Generar el mapa de estaciones
mapa_estaciones()
}
}
